Inhaler: renamed wearable_ble to inhaler_ble
