+++ stm32-smart-inhaler/Core/Src/inhaler_ble.c
@@ -1,4 +1,4 @@
-#include "wearable_ble.h"
+#include "inhaler_ble.h"
 #include "common_blesvc.h"
 #include "app_ble.h"
 #include "pvd.h"
